requirements.txt
pytest==8.3.4
pytest-xdist==3.6.1
pytest-asyncio==0.24.0
allure-pytest==2.13.5
requests==2.32.3
pydantic==2.9.2
ruff==0.6.9
